requirements.txt
aiohttp==3.4.4
pyzmq==17.1.2
PyOTA==2.0.6